Update zendurebattery.py
Allow float capacities for batteries

+++ custom_components/zendure_ha/zendurebattery.py
@@ -19,7 +19,7 @@ class ZendureBattery(ZendureBase):
 
     batterydict: dict[str, ZendureBattery] = {}
 
-    def __init__(self, hass: HomeAssistant, name: str, model: str, snNumber: str, parent: str, kwh: int) -> None:
+    def __init__(self, hass: HomeAssistant, name: str, model: str, snNumber: str, parent: str, kwh: float) -> None:
         """Initialize ZendureBattery."""
         super().__init__(hass, name, model, snNumber, parent)
         self.batterydict[snNumber] = self
